refactor: update todo task list
