+++ frontend/src/components/TodoTaskList/TodoTaskListFooter/TodoTaskListFooterProps.ts
@@ -1,4 +1,4 @@
 export type TodoTaskListFooterProps = {
   completedTaskVisible: boolean;
-  handleCompleteTasks: () => void;
+  toggleTheVisibilityOfDoneTasks: () => void;
 };
