+++ frontend/src/components/TodoTaskList/TodoTaskListFooter/index.styled.tsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 
-export const StyledTodoTaskListFooter = styled.div`
+export const StyledTodoTaskListFooter = styled.footer`
   font-weight: 600;
   font-size: 14px;
   text-transform: uppercase;
